First small refactor to move non-io classes outside of io
