+++ lib/archive.dart
@@ -4,6 +4,7 @@ export 'src/bzip2/bz2_bit_reader.dart';
 export 'src/bzip2/bz2_bit_writer.dart';
 export 'src/bzip2/bzip2.dart';
 export 'src/tar/tar_file.dart';
+export 'src/util/abstract_file_handle.dart';
 export 'src/util/adler32.dart';
 export 'src/util/aes.dart';
 export 'src/util/archive_exception.dart';
@@ -13,6 +14,7 @@ export 'src/util/crc64.dart';
 export 'src/util/input_stream.dart';
 export 'src/util/mem_ptr.dart';
 export 'src/util/output_stream.dart';
+export 'src/util/ram_file_handle.dart';
 export 'src/zip/zip_directory.dart';
 export 'src/zip/zip_file_header.dart';
 export 'src/zip/zip_file.dart';
