+++ lib/src/io/file_handle.dart
@@ -1,25 +1,7 @@
 import 'dart:io';
 import 'dart:typed_data';
 
-enum AbstractFileOpenMode {
-  read,
-  write,
-}
-
-abstract class AbstractFileHandle {
-  AbstractFileOpenMode openMode;
-
-  AbstractFileHandle(this.openMode);
-
-  int get position;
-  set position(int p);
-  int get length;
-  bool get isOpen;
-  Future<void> close();
-  void closeSync();
-  int readInto(Uint8List buffer, [int? end]);
-  void writeFromSync(List<int> buffer, [int start = 0, int? end]);
-}
+import 'package:archive/src/util/abstract_file_handle.dart';
 
 class FileHandle extends AbstractFileHandle {
   final String _path;
