+++ lib/src/io/input_file_stream.dart
@@ -3,7 +3,8 @@ import 'dart:typed_data';
 
 import 'file_buffer.dart';
 import 'file_handle.dart';
-import 'ram_file_handle.dart';
+import '../util/ram_file_handle.dart';
+import '../util/abstract_file_handle.dart';
 import '../util/archive_exception.dart';
 import '../util/byte_order.dart';
 import '../util/input_stream.dart';
