+++ lib/src/io/output_file_stream.dart
@@ -1,8 +1,9 @@
 import 'dart:typed_data';
 
 import 'package:archive/src/io/file_handle.dart';
-import 'package:archive/src/io/ram_file_handle.dart';
+import 'package:archive/src/util/ram_file_handle.dart';
 
+import '../util/abstract_file_handle.dart';
 import '../util/byte_order.dart';
 import '../util/input_stream.dart';
 import '../util/output_stream.dart';
